Splitting the Archive API calls to a separate file.
